Created Laravel 7 Project
Version of Laravel used is 7.x at this moment.

Besides the usual Laravel configurations, I have also added in MongoDB integration into config and serviceprovider files.

With respect to usage of MongoDB in the future, we will be proceeding with Jenssegers MongoDB package version 3.7 that is available for us.

Lastly, I have also started on the basic layout and logic of the authentication system. Front end preset has been set to ReactJS. At this moment, login system allows for creation of new users and allows existing users to log in. However the forgot password or reset password features have not been enabled as I have not added in the service providers. But the logic is there already with respect to Jenssegers package.